requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
biopython==1.83
tqdm==4.65.0
numpy==1.26.0
pandas==2.0.3
